requirements.txt
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.2.0
beautifulsoup4==4.9.2
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-elasticsearch-dsl==7.1.4
django-redis==4.12.1
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
idna==2.10
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
numpy==1.19.2
pandas==1.1.2
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
